Reset latestUvRate on redeem
